requirements.txt
git+https://github.com/NVIDIA/dllogger
nibabel==3.2.1
scipy==1.6.3
torchio==0.18.39
tensorboard==2.5.0
joblib==1.0.1
pytorch-lightning==1.3.8
scikit-learn==1.0
scikit-image==0.18.3
pynvml==11.0.0